@@ -58,32 +58,31 @@
     "rxjs": "catalog:"
   },
   "peerDependencies": {
-    "@angular/core": ">=19.0.0"
+    "@angular/core": ">=20.0.0"
   },
   "devDependencies": {
     "@4tw/cypress-drag-drop": "catalog:",
     "@analogjs/platform": "^2.2.2",
     "@analogjs/vite-plugin-angular": "^2.2.2",
     "@analogjs/vitest-angular": "^2.2.2",
-    "@angular-devkit/build-angular": "^19.2.19",
     "@angular-eslint/eslint-plugin": "catalog:",
-    "@angular/animations": "^19.2.18",
-    "@angular/build": "19.2.19",
-    "@angular/cli": "^19.2.19",
-    "@angular/common": "^19.2.18",
-    "@angular/compiler": "^19.2.18",
-    "@angular/compiler-cli": "^19.2.18",
-    "@angular/core": "^19.2.18",
-    "@angular/forms": "^19.2.18",
-    "@angular/language-service": "^19.2.18",
-    "@angular/platform-browser": "^19.2.18",
-    "@angular/platform-browser-dynamic": "^19.2.18",
-    "@angular/router": "^19.2.18",
+    "@angular/animations": "^20.3.16",
+    "@angular/build": "^20.3.15",
+    "@angular/cli": "^20.3.15",
+    "@angular/common": "^20.3.16",
+    "@angular/compiler": "^20.3.16",
+    "@angular/compiler-cli": "^20.3.16",
+    "@angular/core": "^20.3.16",
+    "@angular/forms": "^20.3.16",
+    "@angular/language-service": "^20.3.16",
+    "@angular/platform-browser": "^20.3.16",
+    "@angular/platform-browser-dynamic": "^20.3.16",
+    "@angular/router": "^20.3.16",
     "@faker-js/faker": "catalog:",
     "@fnando/sparkline": "catalog:",
     "@formkit/tempo": "catalog:",
     "@gc-utils/fs-extra": "^0.4.0",
-    "@ng-select/ng-select": "^14.9.0",
+    "@ng-select/ng-select": "^20.7.0",
     "@ngx-translate/http-loader": "^16.0.1",
     "@nx/angular": "^21.6.10",
     "@nx/devkit": "^21.6.10",
@@ -109,16 +108,16 @@
     "jsdom": "catalog:",
     "jsdom-global": "catalog:",
     "native-copyfiles": "catalog:",
-    "ng-packagr": "^19.2.2",
-    "ngx-bootstrap": "^19.0.2",
+    "ng-packagr": "^20.3.2",
+    "ngx-bootstrap": "^20.0.2",
     "oxlint": "catalog:",
     "remove-glob": "catalog:",
     "rxjs": "catalog:",
     "sass": "catalog:",
     "sortablejs": "catalog:",
     "tslib": "catalog:",
-    "typescript": "~5.8.3",
-    "vite": "catalog:vite6",
+    "typescript": "~5.9.3",
+    "vite": "catalog:",
     "vitest": "catalog:",
     "yaml": "^2.8.2",
     "zone.js": "~0.16.0"
