@@ -1,4 +1,3 @@
-import { NgIf } from '@angular/common';
 import { Component, type OnInit } from '@angular/core';
 import { AngularSlickgridModule, Formatters, type AngularGridInstance, type Column, type GridOption } from '../../library';
 
@@ -6,9 +5,11 @@ const NB_ITEMS = 995;
 
 @Component({
   template: `<button (click)="clickMe()">I'm a button from an Angular component (click me)</button>
-    <div *ngIf="clickedTimes">You've clicked me {{ clickedTimes }} time(s)</div>`,
+    @if (clickedTimes) {
+      <div>You've clicked me {{ clickedTimes }} time(s)</div>
+    }`,
   selector: 'custom-footer',
-  imports: [NgIf],
+  imports: [],
 })
 export class CustomFooterComponent {
   clickedTimes = 0;
