@@ -1,4 +1,3 @@
-import { NgIf } from '@angular/common';
 import { ChangeDetectorRef, Component, type OnInit } from '@angular/core';
 import {
   AngularSlickgridModule,
@@ -13,7 +12,7 @@ import {
 @Component({
   templateUrl: './example10.component.html',
   styles: ['.alert { padding: 8px; margin-bottom: 10px }', '.col-sm-1{ max-width: 70px }'],
-  imports: [AngularSlickgridModule, NgIf],
+  imports: [AngularSlickgridModule],
 })
 export class Example10Component implements OnInit {
   angularGrid1!: AngularGridInstance;
